requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
orjson
motor
beanie
pytest
httpx
requests
celery
redis
flower
emb_opt==1.0.0